feat: implement input capture keybind system
