+++ crates/overlay/src/app.rs
@@ -100,16 +100,17 @@ async fn run_client(mut client: IpcClientConn) -> anyhow::Result<()> {
                 )?;
             }
 
-            Request::SetInputCapture(cmd) => {
-                let applied = Backends::with_backend(
-                    HWND(ptr::null_mut::<c_void>().with_addr(cmd.hwnd as usize)),
-                    |backend| {
-                        backend.capture_input = cmd.capture;
-                    },
-                )
-                .is_some();
-
-                client.reply(id, applied)?;
+            Request::SetInputCaptureKeybind(cmd) => {
+                client.reply(
+                    id,
+                    Backends::with_backend(
+                        HWND(ptr::null_mut::<c_void>().with_addr(cmd.hwnd as usize)),
+                        |backend| {
+                            backend.set_input_capture_keybind(cmd.keybind);
+                        },
+                    )
+                    .is_some(),
+                )?;
             }
 
             Request::UpdateSharedHandle(shared) => {
